+++ src/config/StrictModeDropable.js
@@ -1,0 +1,20 @@
+// This File is a fix to the react-beautiful-dnd library to work with StrictMode
+import { useEffect, useState } from 'react';
+import { Droppable } from 'react-beautiful-dnd';
+
+const StrictModeDroppable = ({ children, ...props }) => {
+  const [enabled, setEnabled] = useState(false);
+  useEffect(() => {
+    const animation = requestAnimationFrame(() => setEnabled(true));
+    return () => {
+      cancelAnimationFrame(animation);
+      setEnabled(false);
+    };
+  }, []);
+  if (!enabled) {
+    return null;
+  }
+  return <Droppable {...props}>{children}</Droppable>;
+};
+
+export default StrictModeDroppable;
